feat(visualizations): port WorkflowVisualization (full lifecycle)
WorkflowVisualization renders the full 42-min cell lifecycle report from
a single trajectory. reports/workflow_report.py keeps the step-based
pipeline orchestration (parca → cache → composite → run → division →
daughters) and dispatches to the Step for rendering. Matplotlib
lazy-imported inside the rendering functions.

This completes the per-Step ports (Tasks 2-8). All 7 Visualization Steps
are now registered.
